fix(send): an agent that names no human still records which agent it was
Codex caught the consequence of my previous fix, and it was the original bug
coming back through the side I had just closed.

Making a passport-less agent store NOTHING left a new row indistinguishable
from a pre-0258 one. The fire path reads a NULL actor id as "this row cannot
say which agent it was" and falls back to the derived `agent:<human-uuid>` —
so the invented identity returned for exactly the actor whose real id had been
in hand at schedule time.

The mistake was upstream of both attempts: the CHECK required the actor id and
the human to travel together, which forced a choice between fabricating a human
and discarding a known agent. Neither is acceptable, and neither had to be.

agent_actor_id is now the anchor — the one fact the fire path cannot
reconstruct — and the passport and the human are independently optional beside
it. Each NULL then means something specific rather than "missing": no passport
means none was presented, no human means the actor named none. What the
constraint still forbids is provenance with no actor, and any agent provenance
on a human-scheduled row.

The all-NULL arm now covers only the rows that already exist, which is what
makes the fire path's reading of a NULL actor id unambiguous: after this
migration, an agent row always names its agent.

TestAnAgentWithNoHumanBehindItStillRecordsWhichAgent asserts both halves — the
row records the agent and no human, and the fired message is audited under that
agent rather than an invented one. Mutation-checked: restore the store-nothing
behaviour and it fails.

The scheduled-send integration file crossed the line ceiling again with the
second assertion, so both agent-provenance tests and their helpers move to
their own file. They are their own concern: not whether a deferred message
fires, but whether the audit can name the actor that produced it.
