+++ backend/internal/compose/integration/scheduledsend_integration_test.go
@@ -443,163 +443,6 @@ func (p *preflightEnv) alarmsFor(t *testing.T, id ids.UUID) int {
 	return n
 }
 
-// scheduleAsAgent defers a message with an agent as the actor, through the same
-// SendOrSchedule the tool surface calls.
-func (p *preflightEnv) scheduleAsAgent(t *testing.T, actor principal.Principal, at time.Time) ids.UUID {
-	t.Helper()
-	anchor, err := ids.Parse(p.activityID)
-	if err != nil {
-		t.Fatalf("the fixture activity id does not parse: %v", err)
-	}
-	out, err := compose.ScheduleAsAgentForTest(context.Background(), p.Pool, p.workspaceID(t), actor,
-		ids.From[ids.ActivityKind](anchor),
-		activities.SendEmailInput{
-			Recipients:     []string{"buyer@preflight.test"},
-			Subject:        "Monday morning",
-			Body:           "Written the night before, by a tool.",
-			ConsentPurpose: "transactional",
-		}, at)
-	if err != nil {
-		t.Fatalf("scheduling as an agent: %v", err)
-	}
-	if out.Scheduled == nil {
-		t.Fatal("scheduling as an agent sent immediately instead of deferring")
-	}
-	return out.Scheduled.ID
-}
-
-// auditActor reads the actor an audit row names for one record.
-func (p *preflightEnv) auditActor(t *testing.T, entityType string, entityID ids.UUID, action string) (string, string) {
-	t.Helper()
-	var actorType, actorID string
-	if err := apptest.InWorkspace(p.AppEnv, t, p.Slug, func(tx pgx.Tx) error {
-		return tx.QueryRow(context.Background(),
-			`SELECT actor_type, actor_id FROM audit_log
-			  WHERE entity_type = $1 AND entity_id = $2 AND action = $3
-			  ORDER BY occurred_at DESC LIMIT 1`,
-			entityType, entityID, action).Scan(&actorType, &actorID)
-	}); err != nil {
-		t.Fatalf("reading the %s audit row for %s: %v", action, entityID, err)
-	}
-	return actorType, actorID
-}
-
-// An agent-scheduled message fires as the agent that scheduled it — by name.
-//
-// The human's id is in the row already, and rebuilding an agent identity from it
-// produces `agent:<human-uuid>`: an actor that never existed, and the same one
-// for every agent and every passport acting for that person. The release audit
-// row, the activity's captured_by and the outbox envelope then cannot say which
-// agent produced the message, which is the attribution ADR-0055 rests on
-// (#1258).
-//
-// What must NOT change with it is the ceiling: the grants are still the human's,
-// re-read live at fire, so preserving the identity cannot widen what the message
-// may do.
-func TestAnAgentScheduledSendFiresUnderTheAgentThatScheduledIt(t *testing.T) {
-	p := setupPreflight(t)
-	p.connect(t, gmailReadonlyScope, gmailSendScope)
-
-	passport := ids.NewV7()
-	agent := principal.Principal{
-		Type:       principal.PrincipalAgent,
-		ID:         "agent:" + passport.String(),
-		PassportID: passport,
-		UserID:     uuidOf(t, p.user),
-		OnBehalfOf: uuidOf(t, p.user),
-		SeatType:   principal.SeatFull,
-		Permissions: principal.Permissions{
-			RoleKeys: []string{"admin"},
-			Objects: map[string]principal.ObjectGrant{
-				"activity": {Create: true, Read: true, Update: true},
-				"person":   {Create: true, Read: true, Update: true},
-			},
-			RowScope: principal.RowScopeAll,
-		},
-	}
-
-	id := p.scheduleAsAgent(t, agent, time.Now().Add(2*time.Hour))
-	p.makeDue(t, id)
-	p.fire(t, id)
-
-	if !p.sent(t, id) {
-		status, reason := p.scheduledStatus(t, id)
-		t.Fatalf("an agent-scheduled message did not fire: %q/%q", status, reason)
-	}
-
-	// The release audit row is where an investigation looks for who acted.
-	actorType, actorID := p.auditActor(t, "activity", p.releasedActivity(t, id), "create")
-	if actorType != "agent" {
-		t.Errorf("the fired message is audited as %q, want agent — a tool-written message recorded as a human's act", actorType)
-	}
-	if actorID != agent.ID {
-		t.Errorf("the fired message names actor %q, want %q — the audit cannot say which agent produced it",
-			actorID, agent.ID)
-	}
-	// The specific shape of the old defect: an identity derived from the human.
-	if actorID == "agent:"+p.user {
-		t.Error("the actor id was rebuilt from the human's id — every agent acting for this person collapses into one invented actor")
-	}
-}
-
-// An agent that names no human stores no provenance, rather than a guess.
-//
-// `agent_on_behalf_of` means "the human behind this agent". The obvious
-// shortcut when it is absent is to fall back to the principal's UserID, and
-// that is wrong: a passport-less agent principal's UserID is an AGENT's own
-// app_user row (compose/extjobsrun.go selects it `WHERE id = $1 AND is_agent`),
-// so the fallback writes an agent's id into a column a human belongs in. The
-// fire path hands that to actor.OnBehalfOf, which auth.Admit reads to derive
-// seat and RBAC — a fabricated authority, the same class of defect the stored
-// provenance exists to end.
-//
-// Storing nothing is the honest record for an actor that never named a human:
-// the row keeps the pre-0258 behaviour instead of asserting a person who was
-// never involved.
-func TestAnAgentWithNoHumanBehindItStoresNoProvenance(t *testing.T) {
-	p := setupPreflight(t)
-	p.connect(t, gmailReadonlyScope, gmailSendScope)
-
-	// No OnBehalfOf and no passport — the extension-tick shape. UserID is the
-	// agent's OWN row, which is exactly what must not be copied.
-	agent := principal.Principal{
-		Type:     principal.PrincipalAgent,
-		ID:       "agent:extension-tick",
-		UserID:   uuidOf(t, p.user),
-		SeatType: principal.SeatFull,
-		Permissions: principal.Permissions{
-			RoleKeys: []string{"admin"},
-			Objects: map[string]principal.ObjectGrant{
-				"activity": {Create: true, Read: true, Update: true},
-				"person":   {Create: true, Read: true, Update: true},
-			},
-			RowScope: principal.RowScopeAll,
-		},
-	}
-
-	id := p.scheduleAsAgent(t, agent, time.Now().Add(2*time.Hour))
-
-	actorID, passport, behalf := p.storedProvenance(t, id)
-	if actorID != nil || passport != nil || behalf != nil {
-		t.Fatalf("an agent naming no human stored provenance %v/%v/%v — the columns travel together, and a guessed human is worse than none",
-			actorID, passport, behalf)
-	}
-}
-
-// storedProvenance reads the three scheduling-time agent columns.
-func (p *preflightEnv) storedProvenance(t *testing.T, id ids.UUID) (actorID *string, passport, behalf *ids.UUID) {
-	t.Helper()
-	if err := apptest.InWorkspace(p.AppEnv, t, p.Slug, func(tx pgx.Tx) error {
-		return tx.QueryRow(context.Background(),
-			`SELECT agent_actor_id, agent_passport_id, agent_on_behalf_of
-			   FROM scheduled_send WHERE id = $1`, id).Scan(&actorID, &passport, &behalf)
-	}); err != nil {
-		t.Fatalf("reading the stored provenance for %s: %v", id, err)
-	}
-	return actorID, passport, behalf
-}
-
-// rowVersion reads the scheduled row's optimistic-concurrency version.
 func (p *preflightEnv) rowVersion(t *testing.T, id ids.UUID) int64 {
 	t.Helper()
 	var version int64
